make abstract DragonPlugin class
